+++ NEWS.md
@@ -1,4 +1,4 @@
-# gittargets 0.0.4.9000
+# gittargets 0.0.5
 
 * Use `processx::run()` instead of `system2()` in `tar_git_ok()` and set `HOME` to `USERPROFILE` on Windows (#12, @psychelzh).